// task1_moviecatalogue/types/domain.ts
export interface Movie {
  id: number;
  title: string;
  category: string;
  image: string;
  overview: string;
  genre_ids: number[];
  release_date: string;
  runtime: number | null;
}

export interface CastMember {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string | null;
  cast_id: number;
  character: string;
  credit_id: string;
  order: number;
}

export interface CrewMember {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string | null;
  credit_id: string;
  department: string;
  job: string;
}

export interface MovieCredits {
  id: number;
  cast: CastMember[];
  crew: CrewMember[];
}

export interface MovieDetails {
  runtime: number | null;
  // Add other details as needed
}